remove-bg-service/requirements.txt
fastapi
uvicorn
python-multipart
rembg[cpu]
numpy
Pillow
